# infoschema: speedup SchemaByID