<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #52610

Problem Summary:

### What changed and how does it work?
- add a schema id->name map

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

insert P99: 7.8ms -> 4ms; P95: 5.5ms -> 3.8ms
![img_v3_02at_8063c40c-b133-40e9-ba0f-c0627225043g](https://github.com/pingcap/tidb/assets/3312245/7ec0a9e5-3af7-416d-a0bd-b3c276b5fd81)

```
mysql> select count(1) from information_schema.schemata;
+----------+
| count(1) |
+----------+
|     7007 |
+----------+

mysql> create table t(id int);
Query OK, 0 rows affected (0.14 sec)

mysql> insert into t values(1);
```

- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```